Moving types to llms
